requirements.txt
appnope==0.1.0
backcall==0.1.0
boto3==1.13.4
botocore==1.16.4
certifi==2020.4.5.1
chardet==3.0.4
decorator==4.4.2
dnspython==1.16.0
docutils==0.15.2
easyzone==1.2.2
idna==2.9
ipython==7.14.0
ipython-genutils==0.2.0
jedi==0.17.0
jmespath==0.9.5
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
ptyprocess==0.6.0
Pygments==2.6.1
python-dateutil==2.8.1
requests==2.23.0
requests-file==1.5.1
s3transfer==0.3.3
six==1.14.0
tldextract==2.2.2
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.1.9
